feat(RELEASE-1725): auto generate README.md files for tasks/pipelines (#22)
Add readme_generator.sh, which generates README.md files
for tasks and pipelines, as well as a check_readme.sh script and
workflow to check that the README.md files in each task and pipeline
match the output of readme-generator.sh.

Also updated all README.md files using this script.

Signed-off-by: Paul Fitzgerald <pafitzge@redhat.com>